+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] / 2023-07-24
+### Updated
+- Update `ricaun.Nuke` to `1.5.1`
+### Tests
+- Add `Tests` project
+
 ## [1.5.0] / 2023-05-31
 ### Updated
 - Update `ricaun.Nuke` to `1.5.0`
@@ -238,6 +244,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.5.1]: ../../compare/1.5.0...1.5.1
 [1.5.0]: ../../compare/1.4.3...1.5.0
 [1.4.3]: ../../compare/1.4.2...1.4.3
 [1.4.2]: ../../compare/1.4.1...1.4.2